requirements.txt
streamlit
python-dotenv
langchain
langchain-openai
langchain-community
langchain-text-splitters
faiss-cpu
pypdf